cv-generation/pyproject.toml
[project]
name = "cv-generation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "dotenv>=0.9.9",
    "google-genai>=1.59.0",
]
